refactor(store-core): extract intervention handlers into handlers/intervention.ts (audit P2-3)
Move the intervention family (handleMultiQuestionIntervention,
applyInterventionBuilder + SessionInterventionBuilder) out of the
handlers/index.ts god-module into handlers/intervention.ts, behind the barrel's
export * re-export so the public surface is unchanged.

Drop the now-unused SessionIntervention type import (from ../types) and
MAX_SESSION_INTERVENTIONS import (from ../utils) from index.ts — both were only
used by the moved handlers.

Pure move — body byte-identical to origin/main. index.ts 1929 -> 1809 lines.
